move single connection HTTP transport implementation
